requirements.txt
anyio==4.9.0
bitunix @ git+https://github.com/0xCherryBlueZu/bitunix.git@00bab865644aabda17a7136793f7fd7d8088c992
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
pandas_ta==0.3.14b0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
